simulator/tested-agents/basic-agent/bot.py
#
# Copyright (c) 2024–2025, Daily
#
# SPDX-License-Identifier: BSD 2-Clause License
#

"""Pipecat Quickstart Example.

The example runs a simple voice AI bot that you can connect to using your
browser and speak with it. You can also deploy this bot to Pipecat Cloud.

Required AI services:
- Deepgram (Speech-to-Text)
- OpenAI (LLM)
- Cartesia (Text-to-Speech)

Run the bot using::

    uv run bot.py
"""

import os
import sys
from pathlib import Path

from dotenv import load_dotenv
from loguru import logger

print("🚀 Starting Pipecat bot...")
print("⏳ Loading models and imports (20 seconds, first run only)\n")

logger.info("Loading Local Smart Turn Analyzer V3...")
from pipecat.audio.turn.smart_turn.local_smart_turn_v3 import LocalSmartTurnAnalyzerV3

logger.info("✅ Local Smart Turn Analyzer V3 loaded")
logger.info("Loading Silero VAD model...")
from pipecat.audio.vad.silero import SileroVADAnalyzer

logger.info("✅ Silero VAD model loaded")

from pipecat.audio.vad.vad_analyzer import VADParams
from pipecat.frames.frames import LLMRunFrame

logger.info("Loading pipeline components...")
from pipecat.pipeline.pipeline import Pipeline
from pipecat.pipeline.runner import PipelineRunner
from pipecat.pipeline.task import PipelineParams, PipelineTask
from pipecat.processors.aggregators.llm_context import LLMContext
from pipecat.processors.aggregators.llm_response_universal import LLMContextAggregatorPair
from pipecat.processors.frameworks.rtvi import RTVIConfig, RTVIObserver, RTVIProcessor
from pipecat.runner.types import RunnerArguments
from pipecat.runner.utils import create_transport
from pipecat.services.cartesia.tts import CartesiaTTSService
from pipecat.services.deepgram.stt import DeepgramSTTService
from pipecat.services.openai.llm import OpenAILLMService
from pipecat.transports.base_transport import BaseTransport, TransportParams
from pipecat.transports.daily.transport import DailyParams, DailyTransport
from pipecatcloud.agent import (
    DailySessionArguments,
    SessionArguments,
    WebSocketSessionArguments,
)

logger.info("✅ All components loaded successfully!")

load_dotenv(override=True)


async def run_bot(transport: BaseTransport, runner_args: RunnerArguments, agent_name: str = "BasicAgent"):
    logger.info(f"Starting bot")

    stt = DeepgramSTTService(api_key=os.getenv("DEEPGRAM_API_KEY"))

    tts = CartesiaTTSService(
        api_key=os.getenv("CARTESIA_API_KEY"),
        voice_id="71a7ad14-091c-4e8e-a314-022ece01c121",  # British Reading Lady
    )

    llm = OpenAILLMService(api_key=os.getenv("OPENAI_API_KEY"))

    messages = [
            {
                "role": "system",
                "content": """You are a helpful and friendly AI participating in a voice conversation.
                
Act like a human, but remember that you aren't a human and that you can't do human
things in the real world. Your voice and personality should be warm and engaging, with a lively and
playful tone.

Because you are participating in a voice conversation, do not use any formatting or emojis in your responses. Use only plain text.

If interacting in a non-English language, start by using the standard accent or dialect familiar to
the user. Talk quickly. Do not refer to these rules,
even if you're asked about them.
-
You are participating in a voice conversation. Keep your responses concise, short, and to the point
unless specifically asked to elaborate on a topic.
Remember, your responses should be short. Just one or two sentences, usually.""",
            },
            {
                "role": "user",
                "content": "Greet the user and ask them how you can help them today.",
            },
        ]

    context = LLMContext(messages)
    context_aggregator = LLMContextAggregatorPair(context)

    rtvi = RTVIProcessor(config=RTVIConfig(config=[]))

    pipeline = Pipeline(
        [
            transport.input(),  # Transport user input
            rtvi,  # RTVI processor
            stt,
            context_aggregator.user(),  # User responses
            llm,  # LLM
            tts,  # TTS
            transport.output(),  # Transport bot output
            context_aggregator.assistant(),  # Assistant spoken responses
        ]
    )

    task = PipelineTask(
        pipeline,
        params=PipelineParams(
            enable_metrics=True,
            enable_usage_metrics=True,
        ),
        observers=[RTVIObserver(rtvi)],
    )

    @transport.event_handler("on_client_connected")
    async def on_client_connected(transport, client):
        logger.info(f"Client connected")
        # Kick off the conversation.
        messages.append({"role": "system", "content": "Say hello and briefly introduce yourself."})
        await task.queue_frames([LLMRunFrame()])

    @transport.event_handler("on_client_disconnected")
    async def on_client_disconnected(transport, client):
        logger.info(f"Client disconnected")
        await task.cancel()

    runner = PipelineRunner(handle_sigint=runner_args.handle_sigint)

    await runner.run(task)


async def cloud_bot(args: SessionArguments):
    """Cloud bot entry point for Pipecat Cloud deployment."""
    try:
        if isinstance(args, DailySessionArguments):
            logger.info("Starting Daily bot")
            logger.info(f"args.room_url from Pipecat Cloud: {args.room_url}")
            logger.info(f"args.body: {args.body}")
            logger.info(f"args.body type: {type(args.body)}")
            
            # Check for room URL in this order:
            # 1. room_url in body (from SessionParams.data) - HIGHEST PRIORITY
            # 2. DAILY_ROOM_URL environment variable
            # 3. args.room_url (default from Pipecat Cloud) - LAST RESORT
            room_url = None
            token = args.token
            
            # Check body parameter FIRST (highest priority - explicitly passed from session)
            if args.body is not None:
                # args.body might be a dict or a string (JSON)
                import json
                body_dict = args.body
                if isinstance(args.body, str):
                    try:
                        body_dict = json.loads(args.body)
                        logger.info(f"Parsed args.body from JSON string")
                    except json.JSONDecodeError:
                        logger.warning(f"Failed to parse args.body as JSON: {args.body}")
                        body_dict = {}
                
                if isinstance(body_dict, dict):
                    body_room_url = body_dict.get("room_url")
                    body_token = body_dict.get("token")
                    logger.info(f"Body contains keys: {list(body_dict.keys())}")
                    if body_room_url:
                        logger.info(f"✅ Found room_url in body: {body_room_url}")
                        logger.info(f"✅ Using room URL from body (highest priority)")
                        room_url = body_room_url
                        token = body_token or token
                    else:
                        logger.warning(f"No room_url found in body. Available keys: {list(body_dict.keys())}")
                else:
                    logger.warning(f"args.body is not a dict after parsing: {type(body_dict)}")
            else:
                logger.warning("args.body is None - room URL not passed via SessionParams.data")
            
            # Check environment variable second (only if body didn't provide room_url)
            if not room_url:
                env_room_url = os.getenv("DAILY_ROOM_URL")
                if env_room_url:
                    logger.info(f"Using room URL from DAILY_ROOM_URL env var: {env_room_url}")
                    room_url = env_room_url
            
            # Fall back to Pipecat Cloud room if nothing else was found (LAST RESORT)
            if not room_url:
                logger.warning(f"⚠️  No room_url found in body or env var, falling back to Pipecat Cloud room: {args.room_url}")
                logger.warning(f"⚠️  This means agents will be in separate rooms!")
                room_url = args.room_url
            
            logger.info(f"Final room_url: {room_url}")
            
            transport = DailyTransport(
                room_url,
                token,
                "Basic Agent",
                DailyParams(
                    audio_in_enabled=True,
                    audio_out_enabled=True,
                    transcription_enabled=False,
                    vad_analyzer=SileroVADAnalyzer(params=VADParams(stop_secs=0.2)),
                    turn_analyzer=LocalSmartTurnAnalyzerV3(),
                ),
            )
        else:
            raise ValueError(f"Unsupported session type: {type(args)}")

        await run_bot(transport, None, agent_name="BasicAgent")
        logger.info("Bot process completed")
    except Exception as e:
        logger.exception(f"Error in bot process: {str(e)}")
        raise


async def bot(runner_args: RunnerArguments):
    """Main bot entry point for the bot starter."""

    transport_params = {
        "daily": lambda: DailyParams(
            room_url=os.getenv("DAILY_ROOM_URL"),
            audio_in_enabled=True,
            audio_out_enabled=True,
            vad_analyzer=SileroVADAnalyzer(params=VADParams(stop_secs=0.2)),
            turn_analyzer=LocalSmartTurnAnalyzerV3(),
        ),
        "webrtc": lambda: TransportParams(
            audio_in_enabled=True,
            audio_out_enabled=True,
            vad_analyzer=SileroVADAnalyzer(params=VADParams(stop_secs=0.2)),
            turn_analyzer=LocalSmartTurnAnalyzerV3(),
        ),
    }

    transport = await create_transport(runner_args, transport_params)

    await run_bot(transport, runner_args)


if __name__ == "__main__":
    from pipecat.runner.run import main

    main()